requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
PyMuPDF==1.24.5
gunicorn==22.0.0
